fix: worksheet auto generate flow
